# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiError": {
        "properties": {
          "code": {
            "description": "Stable machine-readable code.",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "type": "object"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "details"
        ],
        "type": "object"
      },
      "ApiErrorEnvelope": {
        "properties": {
          "data": {
            "type": "null"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object"
          }
        },
        "required": [
          "data",
          "meta",
          "error"
        ],
        "type": "object"
      },
...
      "CreatePresignedUploadDto": {
        "properties": {
          "contentType": {
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp"
            ],
            "type": "string"
          },
          "fileName": {
            "maxLength": 255,
            "type": "string"
          },
          "listingId": {
            "maxLength": 80,
            "type": "string"
          },
          "purpose": {
            "enum": [
              "listing_photo",
              "profile_photo"
            ],
            "type": "string"
          }
        },
        "required": [
          "purpose",
          "fileName",
          "contentType"
        ],
        "type": "object"
      },
...
      "ReorderMediaDto": {
        "properties": {
          "displayOrder": {
            "maximum": 10000,
            "minimum": 0,
            "type": "number"
          }
        },
        "required": [
          "displayOrder"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "Supabase JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "metricsToken": {
        "description": "Dedicated operations metrics token; not a Supabase token.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Stable backend-only REST contracts. All ordinary JSON responses use the MediCN data/meta/error envelope. Socket.IO is documented separately.",
    "title": "MediCN backend API",
    "version": "5.9C"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/listings/{listingId}/photos/{photoId}": {
      "delete": {
        "operationId": "MediaController_deleteListingPhoto",
        "parameters": [
          {
            "in": "path",
            "name": "listingId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "photoId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 400 in the stable error envelope."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 401 in the stable error envelope."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 403 in the stable error envelope."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 404 in the stable error envelope."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 413 in the stable error envelope."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 500 in the stable error envelope."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Media"
        ]
      }
    },
    "/api/v1/listings/{listingId}/photos/{photoId}/order": {
      "patch": {
        "operationId": "MediaController_reorder",
        "parameters": [
          {
            "in": "path",
            "name": "listingId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "photoId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderMediaDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 400 in the stable error envelope."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 401 in the stable error envelope."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 403 in the stable error envelope."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 404 in the stable error envelope."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 413 in the stable error envelope."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 500 in the stable error envelope."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Media"
        ]
      }
    },
...
    "/api/v1/uploads/{assetId}": {
      "delete": {
        "operationId": "UploadsController_remove",
        "parameters": [
          {
            "in": "path",
            "name": "assetId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 400 in the stable error envelope."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 401 in the stable error envelope."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 403 in the stable error envelope."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 404 in the stable error envelope."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 413 in the stable error envelope."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 500 in the stable error envelope."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Uploads"
        ]
      }
    },
    "/api/v1/uploads/{intentId}": {
      "get": {
        "operationId": "UploadsController_status",
        "parameters": [
          {
            "in": "path",
            "name": "intentId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 400 in the stable error envelope."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 401 in the stable error envelope."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 403 in the stable error envelope."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 404 in the stable error envelope."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 413 in the stable error envelope."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 500 in the stable error envelope."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Uploads"
        ]
      }
    },
    "/api/v1/uploads/{intentId}/complete": {
      "post": {
        "operationId": "UploadsController_complete",
        "parameters": [
          {
            "in": "path",
            "name": "intentId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 400 in the stable error envelope."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 401 in the stable error envelope."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 403 in the stable error envelope."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 404 in the stable error envelope."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 413 in the stable error envelope."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 500 in the stable error envelope."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Uploads"
        ]
      }
    },
    "/api/v1/uploads/presigned-url": {
      "post": {
        "operationId": "UploadsController_createPresignedUploadUrl",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePresignedUploadDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 400 in the stable error envelope."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 401 in the stable error envelope."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 403 in the stable error envelope."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 404 in the stable error envelope."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 413 in the stable error envelope."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 429 in the stable error envelope."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 500 in the stable error envelope."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 503 in the stable error envelope."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Uploads"
        ]
      }
    },
...
    "/api/v1/users/me/profile-photo": {
      "delete": {
        "operationId": "MediaController_removeProfile",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 400 in the stable error envelope."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 401 in the stable error envelope."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 403 in the stable error envelope."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 404 in the stable error envelope."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 413 in the stable error envelope."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 500 in the stable error envelope."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Media"
        ]
      }
    },
...
  },
  "servers": [],
...
}